Migrate user_get tests
