+++ libparsec/crates/protocol/tests/user.rs
@@ -10,123 +10,6 @@ use libparsec_tests_fixtures::prelude::*;
 use libparsec_types::prelude::*;
 
 #[parsec_test]
-fn serde_user_get_req() {
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   cmd: "user_get"
-    //   user_id: "109b68ba5cdf428ea0017fc6bcc04d4a"
-    let raw = hex!(
-        "82a3636d64a8757365725f676574a7757365725f6964d92031303962363862613563646634"
-        "32386561303031376663366263633034643461"
-    );
-
-    let req = authenticated_cmds::user_get::Req {
-        user_id: "109b68ba5cdf428ea0017fc6bcc04d4a".parse().unwrap(),
-    };
-
-    let expected = authenticated_cmds::AnyCmdReq::UserGet(req);
-
-    let data = authenticated_cmds::AnyCmdReq::load(&raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = if let authenticated_cmds::AnyCmdReq::UserGet(data) = data {
-        data.dump().unwrap()
-    } else {
-        unreachable!()
-    };
-
-    let data2 = authenticated_cmds::AnyCmdReq::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
-#[case::ok(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   device_certificates: [hex!("666f6f626172")]
-    //   revoked_user_certificate: hex!("666f6f626172")
-    //   status: "ok"
-    //   trustchain: {
-    //     devices: [hex!("666f6f626172")]
-    //     revoked_users: [hex!("666f6f626172")]
-    //     users: [hex!("666f6f626172")]
-    //   }
-    //   user_certificate: hex!("666f6f626172")
-    &hex!(
-        "85b36465766963655f63657274696669636174657391c406666f6f626172b87265766f6b65"
-        "645f757365725f6365727469666963617465c406666f6f626172a6737461747573a26f6baa"
-        "7472757374636861696e83a76465766963657391c406666f6f626172ad7265766f6b65645f"
-        "757365727391c406666f6f626172a5757365727391c406666f6f626172b0757365725f6365"
-        "727469666963617465c406666f6f626172"
-    ),
-    authenticated_cmds::user_get::Rep::Ok {
-        user_certificate: b"foobar".as_ref().into(),
-        revoked_user_certificate: Some(b"foobar".as_ref().into()),
-        device_certificates: vec![b"foobar".as_ref().into()],
-        trustchain: authenticated_cmds::user_get::Trustchain {
-            users: vec![b"foobar".as_ref().into()],
-            devices: vec![b"foobar".as_ref().into()],
-            revoked_users: vec![b"foobar".as_ref().into()],
-        },
-    }
-)]
-#[case::ok_null_revoked_user_cert(
-    // Generated from Rust implementation (Parsec v2.13.0-rc1+dev)
-    // Content:
-    //   device_certificates: [hex!("666f6f626172")]
-    //   revoked_user_certificate: None
-    //   status: "ok"
-    //   trustchain: {
-    //     devices: [hex!("666f6f626172")]
-    //     revoked_users: [hex!("666f6f626172")]
-    //     users: [hex!("666f6f626172")]
-    //   }
-    //   user_certificate: hex!("666f6f626172")
-    //
-    &hex!(
-        "85b36465766963655f63657274696669636174657391c406666f6f626172b87265766f6b65"
-        "645f757365725f6365727469666963617465c0a6737461747573a26f6baa74727573746368"
-        "61696e83a76465766963657391c406666f6f626172ad7265766f6b65645f757365727391c4"
-        "06666f6f626172a5757365727391c406666f6f626172b0757365725f636572746966696361"
-        "7465c406666f6f626172"
-    ),
-    authenticated_cmds::user_get::Rep::Ok {
-        user_certificate: b"foobar".as_ref().into(),
-        revoked_user_certificate: None,
-        device_certificates: vec![b"foobar".as_ref().into()],
-        trustchain: authenticated_cmds::user_get::Trustchain {
-            users: vec![b"foobar".as_ref().into()],
-            devices: vec![b"foobar".as_ref().into()],
-            revoked_users: vec![b"foobar".as_ref().into()],
-        },
-    }
-)]
-#[case::not_found(
-    // Generated from Python implementation (Parsec v2.6.0+dev)
-    // Content:
-    //   status: "not_found"
-    &hex!(
-        "81a6737461747573a96e6f745f666f756e64"
-    ),
-    authenticated_cmds::user_get::Rep::NotFound
-)]
-fn serde_user_get_rep(#[case] raw: &[u8], #[case] expected: authenticated_cmds::user_get::Rep) {
-    let data = authenticated_cmds::user_get::Rep::load(raw).unwrap();
-
-    assert_eq!(data, expected);
-
-    // Also test serialization round trip
-    let raw2 = data.dump().unwrap();
-
-    let data2 = authenticated_cmds::user_get::Rep::load(&raw2).unwrap();
-
-    assert_eq!(data2, expected);
-}
-
-#[parsec_test]
 fn serde_user_create_req() {
     // Generated from Python implementation (Parsec v2.6.0+dev)
     // Content:
